break out repair controller (#194)
add repairs controller
